refactor(imports): Update errors directory structure, and some import
