+++ src/errors/file-errors/index.ts
