+++ src/errors/index.ts
@@ -1,4 +1,4 @@
-export * from './DirErrors';
-export * from './FetchData';
-export * from './FileErrors';
-export * from './NoRuleFound';
+export * from './dir-errors';
+export * from './fetch-data-errors';
+export * from './file-errors';
+export * from './no-rule-found';
