+++ src/errors/no-rule-found/index.ts
