+++ src/rules/DEV/JS-TS/NODE/node-version/index.ts
@@ -1,10 +1,10 @@
-import { WriteFileError } from './../../../../../errors/FileErrors';
+import { WriteFileError } from '../../../../../errors/file-errors';
 import { Ok } from '../../../../../choice';
 import * as cp from 'child_process';
 import TypeScript from '../../../../../stacks/typescript';
 import Node from '../../../../../stacks/node';
 import axios from 'axios';
-import { ReadFileError } from '../../../../../errors/FileErrors';
+import { ReadFileError } from '../../../../../errors/file-errors';
 import { readJSON, writeJSON } from 'fs-extra';
 import { logger } from '../../../../../logger';
 import { Register } from '../../../../../register';
