+++ src/stacks/typescript/index.ts
@@ -1,6 +1,6 @@
 import Globals from '../../utils/globals';
 import { pathExistsInJSON } from '../../utils/json';
-import { ReadFileError } from '../../errors/FileErrors';
+import { ReadFileError } from '../../errors/file-errors';
 import { Register } from '../../register';
 
 @Register.stack
