+++ src/stacks/vue-js/index.ts
@@ -2,7 +2,7 @@ import { Register } from './../../register/index';
 import Globals from '../../utils/globals';
 import { pathExistsInJSON } from '../../utils/json/index';
 import { getExactSemver } from '../../utils/semver/index';
-import { ReadFileError } from '../../errors/FileErrors';
+import { ReadFileError } from '../../errors/file-errors';
 
 @Register.stack
 export default class VueJS {
